persist events as part of the TopicConusmer instead of the Database adapter
