+++ mulighetsrommet-arena-adapter/src/main/kotlin/no/nav/mulighetsrommet/arena/adapter/Database.kt
@@ -3,10 +3,8 @@ package no.nav.mulighetsrommet.arena.adapter
 import com.zaxxer.hikari.HikariConfig
 import com.zaxxer.hikari.HikariDataSource
 import kotliquery.Session
-import kotliquery.queryOf
 import kotliquery.sessionOf
 import org.flywaydb.core.Flyway
-import org.intellij.lang.annotations.Language
 import org.slf4j.LoggerFactory
 
 class Database(databaseConfig: DatabaseConfig) {
@@ -34,18 +32,5 @@ class Database(databaseConfig: DatabaseConfig) {
         logger.debug("Start flyway migrations")
         flyway = Flyway.configure().dataSource(jdbcUrl, databaseConfig.user, databaseConfig.password.value).load()
         flyway.migrate()
-    }
-
-    fun persistKafkaEvent(topic: String, key: String, payload: String) {
-        @Language("PostgreSQL")
-        val query = """
-            insert into events(topic, key, payload)
-            values (?, ?, ?::jsonb)
-            on conflict (topic, key)
-            do update set
-                payload = excluded.payload
-        """.trimIndent()
-        session.run(queryOf(query, topic, key, payload).asUpdate)
-        logger.debug("Persisted kafka event: $topic:$key")
     }
 }
