+++ mulighetsrommet-arena-adapter/src/main/kotlin/no/nav/mulighetsrommet/arena/adapter/consumers/TiltakEndretConsumer.kt
@@ -10,6 +10,7 @@ import no.nav.mulighetsrommet.arena.adapter.MulighetsrommetApiClient
 import no.nav.mulighetsrommet.arena.adapter.utils.ProcessingUtils
 import no.nav.mulighetsrommet.domain.adapter.AdapterTiltak
 import no.nav.mulighetsrommet.domain.arena.ArenaTiltak
+import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
 class TiltakEndretConsumer(
@@ -18,7 +19,7 @@ class TiltakEndretConsumer(
     private val client: MulighetsrommetApiClient
 ) : TopicConsumer<ArenaTiltak>(db) {
 
-    private val logger = LoggerFactory.getLogger(TiltakEndretConsumer::class.java)
+    override val logger: Logger = LoggerFactory.getLogger(TiltakEndretConsumer::class.java)
 
     override fun toDomain(payload: JsonElement): ArenaTiltak {
         return Json.decodeFromJsonElement(payload.jsonObject["after"]!!)
