+++ mulighetsrommet-arena-adapter/src/main/kotlin/no/nav/mulighetsrommet/arena/adapter/consumers/TiltakdeltakerEndretConsumer.kt
@@ -10,6 +10,7 @@ import no.nav.mulighetsrommet.arena.adapter.MulighetsrommetApiClient
 import no.nav.mulighetsrommet.arena.adapter.utils.ProcessingUtils
 import no.nav.mulighetsrommet.domain.adapter.AdapterTiltakdeltaker
 import no.nav.mulighetsrommet.domain.arena.ArenaTiltakdeltaker
+import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
 class TiltakdeltakerEndretConsumer(
@@ -18,7 +19,7 @@ class TiltakdeltakerEndretConsumer(
     private val client: MulighetsrommetApiClient
 ) : TopicConsumer<ArenaTiltakdeltaker>(db) {
 
-    private val logger = LoggerFactory.getLogger(TiltakdeltakerEndretConsumer::class.java)
+    override val logger: Logger = LoggerFactory.getLogger(TiltakdeltakerEndretConsumer::class.java)
 
     override fun toDomain(payload: JsonElement): ArenaTiltakdeltaker {
         return Json.decodeFromJsonElement(payload.jsonObject["after"]!!)
