+++ mulighetsrommet-arena-adapter/src/main/kotlin/no/nav/mulighetsrommet/arena/adapter/consumers/TopicConsumer.kt
@@ -2,10 +2,14 @@ package no.nav.mulighetsrommet.arena.adapter.consumers
 
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonElement
+import kotliquery.queryOf
 import no.nav.mulighetsrommet.arena.adapter.Database
 import org.apache.kafka.clients.consumer.ConsumerRecord
+import org.intellij.lang.annotations.Language
+import org.slf4j.Logger
 
 abstract class TopicConsumer<T>(private val db: Database) {
+    abstract val logger: Logger
     abstract val topic: String
 
     fun processEvent(event: ConsumerRecord<String, String>) {
@@ -13,12 +17,15 @@ abstract class TopicConsumer<T>(private val db: Database) {
         val decodedPayload = toDomain(eventPayload)
         if (shouldProcessEvent(decodedPayload)) {
             val key = resolveKey(decodedPayload)
-            db.persistKafkaEvent(
+
+            logger.debug("Persisting event: topic=$topic, key=$key")
+            persistEvent(
                 event.topic(),
                 key,
                 event.value()
             )
 
+            logger.debug("Handling event: topic=$topic, key=$key")
             handleEvent(decodedPayload)
         }
     }
@@ -30,4 +37,17 @@ abstract class TopicConsumer<T>(private val db: Database) {
     protected abstract fun resolveKey(payload: T): String
 
     protected abstract fun handleEvent(payload: T)
+
+    private fun persistEvent(topic: String, key: String, payload: String) {
+        @Language("PostgreSQL")
+        val query = """
+            insert into events(topic, key, payload)
+            values (?, ?, ?::jsonb)
+            on conflict (topic, key)
+            do update set
+                payload = excluded.payload
+        """.trimIndent()
+
+        db.session.run(queryOf(query, topic, key, payload).asUpdate)
+    }
 }
